requirements.txt
munch==2.5.0
networkx==2.6.3
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
pandas==1.1.5
Pillow==8.3.1
pretrainedmodels==0.7.4
protobuf==3.17.3
pyparsing==3.0.7
pyquery
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
pyyaml==6.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Shapely==1.8.0
SimpleITK==2.0.2
tensorboardX==2.4
threadpoolctl==3.0.0
thop
tifffile==2021.7.2
timm==0.4.12
torchinfo
tqdm==4.61.2
xlrd==1.2.0
yacs==0.1.8
